Ctrl+double click on a square to copy its coordinates
With these changes you can ctrl+double click anywhere on the map to copy the
coordinates of that tile.
The format of the copied text is |x=10|y=20|plane=0 for example.
This is useful when filling out maps for use on the wiki.
With the restriction to only copying while holding ctrl, this should not
be experienced as obtrusive.