Make use of the copy API instead of the execCommand function
The mdn page I was following didn't mention the execCommand being
deprecated, but mdn does mention this feature being 'experimental',
hence why I chose the former. Apparently the page I used was outdated
though.

With this the only change from before previous commit is to the
leaflet.position.js file, and only to add the event handler for
ctrl+dblclick to copy.